Extract generic functions into library
